requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
openai==1.6.1
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pillow==10.1.0
pytesseract==0.3.10
boto3==1.34.10
numpy==1.26.2
pgvector==0.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
